@@ -1,8 +1,10 @@
 module github.com/circonus-labs/go-apiclient
 
-go 1.15
+go 1.17
 
 require (
 	github.com/hashicorp/go-retryablehttp v0.7.1
 	github.com/pkg/errors v0.9.1
 )
+
+require github.com/hashicorp/go-cleanhttp v0.5.1 // indirect
