@@ -7,7 +7,7 @@ package apiclient
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"reflect"
@@ -43,7 +43,7 @@ func testMaintenanceServer() *httptest.Server {
 				fmt.Fprintln(w, string(ret))
 			case "PUT":
 				defer r.Body.Close()
-				b, err := ioutil.ReadAll(r.Body)
+				b, err := io.ReadAll(r.Body)
 				if err != nil {
 					panic(err)
 				}
@@ -88,7 +88,7 @@ func testMaintenanceServer() *httptest.Server {
 				}
 			case "POST":
 				defer r.Body.Close()
-				_, err := ioutil.ReadAll(r.Body)
+				_, err := io.ReadAll(r.Body)
 				if err != nil {
 					panic(err)
 				}
@@ -141,16 +141,30 @@ func TestFetchMaintenanceWindow(t *testing.T) {
 	apih, server := maintenanceTestBootstrap(t)
 	defer server.Close()
 
-	tests := []struct { //nolint:govet
+	tests := []struct {
 		id           string
 		cid          string
 		expectedType string
-		shouldFail   bool
 		expectedErr  string
+		shouldFail   bool
 	}{
-		{"empty cid", "", "", true, "invalid maintenance window CID (none)"},
-		{"short cid", "1234", "*apiclient.Maintenance", false, ""},
-		{"long cid", "/maintenance/1234", "*apiclient.Maintenance", false, ""},
+		{
+			id:          "empty cid",
+			shouldFail:  true,
+			expectedErr: "invalid maintenance window CID (none)",
+		},
+		{
+			id:           "short cid",
+			cid:          "1234",
+			expectedType: "*apiclient.Maintenance",
+			shouldFail:   false,
+		},
+		{
+			id:           "long cid",
+			cid:          "/maintenance/1234",
+			expectedType: "*apiclient.Maintenance",
+			shouldFail:   false,
+		},
 	}
 
 	for _, test := range tests {
@@ -192,16 +206,30 @@ func TestUpdateMaintenanceWindow(t *testing.T) {
 	apih, server := maintenanceTestBootstrap(t)
 	defer server.Close()
 
-	tests := []struct { //nolint:govet
-		id           string
+	tests := []struct {
 		cfg          *Maintenance
+		id           string
 		expectedType string
-		shouldFail   bool
 		expectedErr  string
+		shouldFail   bool
 	}{
-		{"invalid (nil)", nil, "", true, "invalid maintenance window config (nil)"},
-		{"invalid (cid)", &Maintenance{CID: "/invalid"}, "", true, "invalid maintenance window CID (/invalid)"},
-		{"valid", &testMaintenance, "*apiclient.Maintenance", false, ""},
+		{
+			id:          "invalid (nil)",
+			shouldFail:  true,
+			expectedErr: "invalid maintenance window config (nil)",
+		},
+		{
+			id:          "invalid (cid)",
+			cfg:         &Maintenance{CID: "/invalid"},
+			shouldFail:  true,
+			expectedErr: "invalid maintenance window CID (/invalid)",
+		},
+		{
+			id:           "valid",
+			cfg:          &testMaintenance,
+			expectedType: "*apiclient.Maintenance",
+			shouldFail:   false,
+		},
 	}
 
 	for _, test := range tests {
@@ -229,15 +257,24 @@ func TestCreateMaintenanceWindow(t *testing.T) {
 	apih, server := maintenanceTestBootstrap(t)
 	defer server.Close()
 
-	tests := []struct { //nolint:govet
-		id           string
+	tests := []struct {
 		cfg          *Maintenance
+		id           string
 		expectedType string
-		shouldFail   bool
 		expectedErr  string
+		shouldFail   bool
 	}{
-		{"invalid (nil)", nil, "", true, "invalid maintenance window config (nil)"},
-		{"valid", &testMaintenance, "*apiclient.Maintenance", false, ""},
+		{
+			id:          "invalid (nil)",
+			shouldFail:  true,
+			expectedErr: "invalid maintenance window config (nil)",
+		},
+		{
+			id:           "valid",
+			cfg:          &testMaintenance,
+			expectedType: "*apiclient.Maintenance",
+			shouldFail:   false,
+		},
 	}
 
 	for _, test := range tests {
@@ -265,14 +302,22 @@ func TestDeleteMaintenanceWindow(t *testing.T) {
 	apih, server := maintenanceTestBootstrap(t)
 	defer server.Close()
 
-	tests := []struct { //nolint:govet
-		id          string
+	tests := []struct {
 		cfg         *Maintenance
-		shouldFail  bool
+		id          string
 		expectedErr string
+		shouldFail  bool
 	}{
-		{"invalid (nil)", nil, true, "invalid maintenance window config (nil)"},
-		{"valid", &testMaintenance, false, ""},
+		{
+			id:          "invalid (nil)",
+			shouldFail:  true,
+			expectedErr: "invalid maintenance window config (nil)",
+		},
+		{
+			id:         "valid",
+			cfg:        &testMaintenance,
+			shouldFail: false,
+		},
 	}
 
 	for _, test := range tests {
@@ -300,15 +345,27 @@ func TestDeleteMaintenanceWindowByCID(t *testing.T) {
 	apih, server := maintenanceTestBootstrap(t)
 	defer server.Close()
 
-	tests := []struct { //nolint:govet
+	tests := []struct {
 		id          string
 		cid         string
-		shouldFail  bool
 		expectedErr string
+		shouldFail  bool
 	}{
-		{"empty cid", "", true, "invalid maintenance window CID (none)"},
-		{"short cid", "1234", false, ""},
-		{"long cid", "/maintenance/1234", false, ""},
+		{
+			id:          "empty cid",
+			shouldFail:  true,
+			expectedErr: "invalid maintenance window CID (none)",
+		},
+		{
+			id:         "short cid",
+			cid:        "1234",
+			shouldFail: false,
+		},
+		{
+			id:         "long cid",
+			cid:        "/maintenance/1234",
+			shouldFail: false,
+		},
 	}
 
 	for _, test := range tests {
@@ -340,18 +397,38 @@ func TestSearchMaintenances(t *testing.T) {
 	search := SearchQueryType("/check_bundle/1234")
 	filter := SearchFilterType(map[string][]string{"f_start_gt": {"1483639916"}})
 
-	tests := []struct { //nolint:govet
-		id           string
+	tests := []struct {
 		search       *SearchQueryType
 		filter       *SearchFilterType
+		id           string
 		expectedType string
-		shouldFail   bool
 		expectedErr  string
+		shouldFail   bool
 	}{
-		{"no search, no filter", nil, nil, expectedType, false, ""},
-		{"search no filter", &search, nil, expectedType, false, ""},
-		{"filter no search", nil, &filter, expectedType, false, ""},
-		{"both filter and search", &search, &filter, expectedType, false, ""},
+		{
+			id:           "no search, no filter",
+			expectedType: expectedType,
+			shouldFail:   false,
+		},
+		{
+			id:           "search no filter",
+			search:       &search,
+			expectedType: expectedType,
+			shouldFail:   false,
+		},
+		{
+			id:           "filter no search",
+			filter:       &filter,
+			expectedType: expectedType,
+			shouldFail:   false,
+		},
+		{
+			id:           "both filter and search",
+			search:       &search,
+			filter:       &filter,
+			expectedType: expectedType,
+			shouldFail:   false,
+		},
 	}
 
 	for _, test := range tests {
