@@ -7,7 +7,7 @@ package apiclient
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"reflect"
@@ -47,7 +47,7 @@ func testMetricServer() *httptest.Server {
 				fmt.Fprintln(w, string(ret))
 			case "PUT":
 				defer r.Body.Close()
-				b, err := ioutil.ReadAll(r.Body)
+				b, err := io.ReadAll(r.Body)
 				if err != nil {
 					panic(err)
 				}
@@ -122,16 +122,30 @@ func TestFetchMetric(t *testing.T) {
 	apih, server := metricTestBootstrap(t)
 	defer server.Close()
 
-	tests := []struct { //nolint:govet
+	tests := []struct {
 		id           string
 		cid          string
 		expectedType string
-		shouldFail   bool
 		expectedErr  string
+		shouldFail   bool
 	}{
-		{"empty cid", "", "", true, "invalid metric CID (none)"},
-		{"short cid", "1234_foo", "*apiclient.Metric", false, ""},
-		{"long cid", "/metric/1234_foo", "*apiclient.Metric", false, ""},
+		{
+			id:          "empty cid",
+			shouldFail:  true,
+			expectedErr: "invalid metric CID (none)",
+		},
+		{
+			id:           "short cid",
+			cid:          "1234_foo",
+			expectedType: "*apiclient.Metric",
+			shouldFail:   false,
+		},
+		{
+			id:           "long cid",
+			cid:          "/metric/1234_foo",
+			expectedType: "*apiclient.Metric",
+			shouldFail:   false,
+		},
 	}
 
 	for _, test := range tests {
@@ -173,16 +187,30 @@ func TestUpdateMetric(t *testing.T) {
 	apih, server := metricTestBootstrap(t)
 	defer server.Close()
 
-	tests := []struct { //nolint:govet
-		id           string
+	tests := []struct {
 		cfg          *Metric
+		id           string
 		expectedType string
-		shouldFail   bool
 		expectedErr  string
+		shouldFail   bool
 	}{
-		{"invalid (nil)", nil, "", true, "invalid metric config (nil)"},
-		{"invalid (cid)", &Metric{CID: "/invalid"}, "", true, "invalid metric CID (/invalid)"},
-		{"valid", &testMetric, "*apiclient.Metric", false, ""},
+		{
+			id:          "invalid (nil)",
+			shouldFail:  true,
+			expectedErr: "invalid metric config (nil)",
+		},
+		{
+			id:          "invalid (cid)",
+			cfg:         &Metric{CID: "/invalid"},
+			shouldFail:  true,
+			expectedErr: "invalid metric CID (/invalid)",
+		},
+		{
+			id:           "valid",
+			cfg:          &testMetric,
+			expectedType: "*apiclient.Metric",
+			shouldFail:   false,
+		},
 	}
 
 	for _, test := range tests {
@@ -214,18 +242,38 @@ func TestSearchMetrics(t *testing.T) {
 	search := SearchQueryType("vm`memory`used")
 	filter := SearchFilterType(map[string][]string{"f_tags_has": {"service:cache"}})
 
-	tests := []struct { //nolint:govet
-		id           string
+	tests := []struct {
 		search       *SearchQueryType
 		filter       *SearchFilterType
+		id           string
 		expectedType string
-		shouldFail   bool
 		expectedErr  string
+		shouldFail   bool
 	}{
-		{"no search, no filter", nil, nil, expectedType, false, ""},
-		{"search no filter", &search, nil, expectedType, false, ""},
-		{"filter no search", nil, &filter, expectedType, false, ""},
-		{"both filter and search", &search, &filter, expectedType, false, ""},
+		{
+			id:           "no search, no filter",
+			expectedType: expectedType,
+			shouldFail:   false,
+		},
+		{
+			id:           "search no filter",
+			search:       &search,
+			expectedType: expectedType,
+			shouldFail:   false,
+		},
+		{
+			id:           "filter no search",
+			filter:       &filter,
+			expectedType: expectedType,
+			shouldFail:   false,
+		},
+		{
+			id:           "both filter and search",
+			search:       &search,
+			filter:       &filter,
+			expectedType: expectedType,
+			shouldFail:   false,
+		},
 	}
 
 	for _, test := range tests {
